+++ src/components/Hero/Hero.tsx
@@ -55,9 +55,14 @@ export interface HeroProps extends BoxProps {
    * Set `isDarkText` to `true` if the `textBackgroundColor` is set to a light
    * color. */
   isDarkText?: boolean;
-  /** Optional boolean used to toggle the treatment of the background image in
-   * the "campaign" variant. If true, the background image will be converted to
-   * black & white and darkened to 60% black. */
+  /** Optional boolean used to toggle the blur treatment of the background image
+   * in the "campaign" variant. If true, the background image will be blurred
+   * and darkened slightly.
+   */
+  isBlurredBackgroundImage?: boolean;
+  /** Optional boolean used to toggle the color treatment of the background
+   * image in the "campaign" variant. If true, the background image will be
+   * converted to black & white and darkened moderately. */
   isDarkBackgroundImage?: boolean;
   /** Optional string used for the subheader that displays underneath the
    * heading element. */
@@ -97,6 +102,7 @@ export const Hero: ChakraComponent<
           src: "",
         },
         isDarkText,
+        isBlurredBackgroundImage = false,
         isDarkBackgroundImage = false,
         subHeaderText,
         textBackgroundColor,
@@ -105,6 +111,8 @@ export const Hero: ChakraComponent<
       } = props;
       const styles = useMultiStyleConfig("Hero", {
         foregroundColor,
+        isBlurredBackgroundImage,
+        isDarkBackgroundImage,
         isDarkText,
         textColor,
         variant,
@@ -209,8 +217,9 @@ export const Hero: ChakraComponent<
       } else if (variant === "campaign") {
         /**
          * For better control of the background image in the "campaign" variant,
-         * the image and the associated styles were moved into the `:before`
-         * element.
+         * the background image and the associated styles are handled with a
+         * separate DOM element rather that being applied to the main component
+         * container.
          */
         const campaignBgStyles = {
           content: `""`,
@@ -221,23 +230,47 @@ export const Hero: ChakraComponent<
           top: 0,
           width: "100%",
         };
+        // Style background image based on configuration
+        const finalBackgroundFilters = [
+          // Make grayscale for "dark" style
+          isDarkBackgroundImage && "grayscale(100%)",
+          // Blur image for "blurred" style
+          isBlurredBackgroundImage && "blur(80px)",
+          // Adjust brightness based on style
+          isDarkBackgroundImage
+            ? "brightness(0.4)" // Much darker for "dark" style
+            : isBlurredBackgroundImage
+            ? "brightness(0.8)" // Slightly darker for "blurred" style
+            : null,
+        ]
+          .filter(Boolean)
+          .join(" ");
+
         backgroundImageStyle = backgroundImageSrc
           ? {
+              minHeight: "320px",
+              ".heroBackgroundImage": {
+                ...campaignBgStyles,
+                overflow: "hidden",
+                _after: {
+                  ...campaignBgStyles,
+                  backgroundImage: `/**/url("${backgroundImageSrc}")`,
+                  backgroundPosition: "center",
+                  backgroundSize: "cover",
+                  filter: finalBackgroundFilters,
+                  height: "100%",
+                  overflow: "hidden",
+
+                  // Increase size to hide pixelated edges caused by the blur filter
+                  transform: isBlurredBackgroundImage
+                    ? "scale(1.4)"
+                    : undefined,
+                },
+              },
               _before: {
                 ...campaignBgStyles,
                 bgColor: "ui.black",
-              },
-              _after: {
-                ...campaignBgStyles,
-                backgroundBlendMode: isDarkBackgroundImage
-                  ? "saturation"
-                  : null,
-                backgroundImage: isDarkBackgroundImage
-                  ? `/**/linear-gradient(black, black), url("${backgroundImageSrc}")`
-                  : `/**/url("${backgroundImageSrc}")`,
-                backgroundPosition: "center",
-                backgroundSize: "cover",
-                opacity: isDarkBackgroundImage ? "0.4" : "1.0",
+                overflow: "hidden",
               },
             }
           : backdropBackgroundColor
@@ -346,7 +379,12 @@ export const Hero: ChakraComponent<
             {contentPrep}
           </Box>
         ) : (
-          contentPrep
+          <>
+            {variant === "campaign" && (
+              <Box className="heroBackgroundImage"></Box>
+            )}
+            {contentPrep}
+          </>
         );
 
       return (